Show confirmation messages and loading SweetAlert2

+++ src/modules/daybook/views/EntryView.vue
@@ -46,6 +46,7 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex' // computed!!!, methods!!!
+import Swal from 'sweetalert2'
 
 import getDayMonthYear from '../helpers/getDayMonthYear'
 
@@ -97,6 +98,13 @@ export default {
       this.entry = entry
     },
     async saveEntry() {
+      new Swal({
+        title: 'Espere por favor',
+        allowOutsideClick: false
+      })
+
+      Swal.showLoading()
+
       if ( this.entry.id ) {
         // Actualizar
         await this.updateEntry( this.entry )      
@@ -106,11 +114,29 @@ export default {
 
         this.$router.push({ name: 'entry', params: { id } })
       }
+
+      Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
     },
     async onDeleteEntry() {
-      await this.deleteEntry( this.entry.id )
+      const { isConfirmed } = await Swal.fire({
+        title: '¿Está seguro?',
+        text: 'Una vez borrado, no se puede recuperar',
+        showDenyButton: true,
+        confirmButtonText: 'Sí, estoy seguro'
+      })
 
-      this.$router.push({ name: 'no-entry' })
+      if( isConfirmed ) {
+        new Swal({
+          title: 'Espere por favor',
+          allowOutsideClick: false
+        })
+
+        Swal.showLoading()
+        await this.deleteEntry( this.entry.id )
+        this.$router.push({ name: 'no-entry' })
+
+        Swal.fire('Eliminado', '', 'success')
+      }
     }
   },
   created() {
